refactor(app_view): split window-chrome helpers into chrome.rs
Lifts the 11 window-chrome / overlay helper functions out of
mod.rs into a new `app_view::chrome` child module:

  * `profile_context_menu_overlay` + `profile_menu_item` — the
    sidebar right-click popup and its single row primitive
  * `suspended_banner` + `suspended_pane` — the two visual
    treatments for a session in `Suspend` mode
  * `about_dialog_body` — content for the About dialog
  * `friendly_open_error` — serial-open error string formatter
    (the only non-rendering helper in the file; lives here
    because every caller is a render-time error path)
  * `welcome_pane` — the empty-state shown at boot with no
    connected profile
  * `session_header` + `status_bar` + `session_overflow_button` —
    the top toolbar, bottom status row, and `⋯` overflow menu
  * `sidebar_header` + `sidebar_icon_strip` — the two layouts
    for the left sidebar (expanded vs. collapsed)

All eleven are `pub(super)` so mod.rs imports them by name.
Internal cross-references (e.g. `session_overflow_button` calling
`profile_menu_item`) stay private. The trailing comment block
about chrome consts having migrated to `SkinTokens` rides along
with the chrome code so the historical note keeps its proximity
to the helpers it describes.

While here: rolled `session_overflow_button` (which had stayed
in mod.rs after the earlier Phase D split) into chrome.rs since
it's an inline child of `session_header` and shares the same
popup-paint pattern as `profile_context_menu_overlay`. Cleans up
the `use buttons::{line_pill, ...}` line on mod.rs too —
`line_pill` was only used by `session_header`, which now lives
in chrome.rs.

mod.rs drops by ~990 lines (6,643 → 5,651). chrome.rs lands at
1,032 lines — large but cohesively scoped to "render-time
chrome subtree producers."

Verified: `cargo clippy --all-targets -- -D warnings` and
`cargo test` both pass (63 tests, unchanged).
